[FEATURE] Make clones of APDS9960 work too with this library
There are some clones of the APDS9960 out in the wild which use a
different device id. Per default the library fails to during the device init
if the id is not 0xAB like the original one. But some clones use 0xA8 as an
id and some others 0x9C.

So improve handling of device ids and make those devices also work
with this library.